+++ services/chainstorage/src/main/java/tech/pegasys/teku/services/chainstorage/StorageService.java
@@ -17,12 +17,14 @@ import static tech.pegasys.teku.spec.config.Constants.STORAGE_QUERY_CHANNEL_PARA
 
 import tech.pegasys.teku.ethereum.pow.api.Eth1EventsChannel;
 import tech.pegasys.teku.infrastructure.async.SafeFuture;
+import tech.pegasys.teku.infrastructure.async.eventthread.AsyncRunnerEventThread;
 import tech.pegasys.teku.service.serviceutils.Service;
 import tech.pegasys.teku.service.serviceutils.ServiceConfig;
 import tech.pegasys.teku.storage.api.Eth1DepositStorageChannel;
 import tech.pegasys.teku.storage.api.StorageQueryChannel;
 import tech.pegasys.teku.storage.api.StorageUpdateChannel;
 import tech.pegasys.teku.storage.api.VoteUpdateChannel;
+import tech.pegasys.teku.storage.server.BatchingVoteUpdateChannel;
 import tech.pegasys.teku.storage.server.ChainStorage;
 import tech.pegasys.teku.storage.server.Database;
 import tech.pegasys.teku.storage.server.DepositStorage;
@@ -71,7 +73,12 @@ public class StorageService extends Service implements StorageServiceFacade {
               .subscribe(Eth1DepositStorageChannel.class, depositStorage)
               .subscribe(Eth1EventsChannel.class, depositStorage)
               .subscribe(StorageUpdateChannel.class, chainStorage)
-              .subscribe(VoteUpdateChannel.class, chainStorage, VOTE_UPDATE_QUEUE_CAPACITY)
+              .subscribe(
+                  VoteUpdateChannel.class,
+                  new BatchingVoteUpdateChannel(
+                      chainStorage,
+                      new AsyncRunnerEventThread(
+                          "batch-vote-updater", serviceConfig.getAsyncRunnerFactory())))
               .subscribeMultithreaded(
                   StorageQueryChannel.class, chainStorage, STORAGE_QUERY_CHANNEL_PARALLELISM);
         });